worker.workers refactor into worker.scribe_worker
