+++ horde-scribe-bridge.py
@@ -8,7 +8,7 @@ set_worker_env_vars_from_config()  # Get `cache_home` from `bridgeconfig.yaml` i
 
 from worker.bridge_data import BridgeData  # noqa: E402
 from worker.logger import logger, quiesce_logger, set_logger_verbosity  # noqa: E402
-from worker.workers.scribe import ScribeWorker  # noqa: E402
+from worker.scribe_worker import ScribeWorker  # noqa: E402
 
 
 def main():
@@ -16,6 +16,7 @@ def main():
     quiesce_logger(args.quiet)
     bridge_data = BridgeData()
     bridge_data.reload_data()
+
     try:
         worker = ScribeWorker(bridge_data)
         worker.start()
